添加了appeal 在登陆添加了oss信息

+++ com.yrtech.SurveyAPI/DTO/Appeal/AppealDto.cs
@@ -8,7 +8,7 @@ namespace com.yrtech.SurveyAPI.DTO
 {
     public class AppealDto
     {
-        public int AppealId { get; set; }
+        public long AppealId { get; set; }
         public int ProjectId { get; set; }
         public string ProjectCode { get; set; }
         public string ProjectName { get; set; }
